pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "replayt"
version = "0.1.0"
description = "Deterministic LLM workflows you can replay — explicit states, strict schemas, local logs."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "replayt contributors" }]
keywords = ["llm", "workflow", "fsm", "pydantic", "replay", "deterministic"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Typing :: Typed",
]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.6.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
yaml = ["PyYAML>=6.0"]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
]

[project.scripts]
replayt = "replayt.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
